refractor: Move the components related to Recipe card in a seperate directory
